Object and Space now have both GetComponent and TryGetComponent funcitons.
I've been meaning to add this for a while but only just now got around to it. The idea here is to have one get function where the component is expected to be part of a space member and another where it may or may not be part of a space member. By using GetComponent, Varkor can crash to let us know that our assumption about that component being present is false. If we have no idea whether the component will be present or not, then TryGet is the function to use.