Shader::Compile no longer gives inaccurrate error line information when there is no newline at the end of some shader files.
I forget exactly what conditions resulted in incorrect line information, but I am relatively certain that the line information would be wrong if specific files in the include chain did not end with new lines.

This commit addresses that, but I also took it as an opportunity to clean up the HandleIncludes function with a non-recursive solution. This is easier to both read and debug. I believe I used a recursive solution beforehand for elegance, but at the end of the day, it was harder to debug and understand. It might have been faster too, but does that really matter? This code is just meant for shader files and I don't expect that this code will need to deal with any extremely large files any time soon, if ever. Preoptimization, don't do it.