Shader includes can now be relative to the file or relative to the executable.
This isn't entirely necessary, but the result is some less messy include paths. The main reason it's being added is to handle cases where a shader within a subproject might want to include a shader library file from vres/.

For example, say the following file exists relative to the executable.

res/shader/test.vs

And this to includes the following shader file.

vres/shader/uniformBuffers/example.glsl

test.vs can use that full path for the inclusion rather than the following.

\#include "../../vres/shader/uniformBuffers/example.glsl"

A minor update to the FailedIncludeSubdirectory test was added to make sure this is working.