Removed unnecessary template specializations from Vlk::Value
The template specializations I made for As and operator= weren't necessary. std::stringstream is able to handle all of those conversions for me.

I am removing these now because I will be adding Extensions in a future commit that will allow me to extend the possible types that can be used for serialization and deserialization.