The first Valkor extension type Vec3.
The entire idea here is to extend Valkor so it can handle different types without muddying the interface of Valkor or the type. To do that, there is a file called Extensions where all of the extensions for different types are placed. These are really just template specializations at the end of the day.

This also adds a unit test for the Vec3 extension.

I would eventually like to separate Valkor into a separate project so it can be readily used in other projects without needing the rest of the engine. I tried to keep the extensions as separate as possible to make this task easier for my future self.