Vlk::Explorer errors no longer print out file and line information.
This verbosity wasn't necessary as it didn't help when pointing out what the problem is. Additionally, the line number information has changed multiple times in the Explorer's unit test and it's annoying to continuously update.

Optimally, we should be printing out the name of the function that called the explorer function. This would tell the user exactly what line the error happened on within their code. I don't know exactly how I would go about achieving this though and I have not reached a point where it would be very helpful for me.