Specular reflections are now working again because the view position is set.
Since the view position is very general and closely related to the matrices in the old Matrices uniform buffer, I decided to add it to that buffer and rename it to the universal uniform buffer. This buffer can then serve as a place for general purpose values that are shared across shaders.