Members can now be deleted from the overview interface.
This could have been added ages ago, but I just haven't had a use for it until now. I also added the "click again to deselect" feature to members in the overview interface.
The DisplayMember function now takes in a double pointer to the inspector interface because the inspector can be created or destroyed within the function and therefore needs to update the pointer.