FindInterface will no longer return interfaces that are about to close.
This is a fix to a problem that was introduced by 8293c4a. Check out this line.

https://github.com/Underdisc/Varkor/blob/8293c4ad5fb98286967c19c48955cc4c3de86e5d/src/gfx/Renderer.cc#L138

This uses the FindInterface function to get the OverviewInterface attached to the CoreInterface. Unfortunately, it might be the case that the OverviewInterface was just closed by the CoreInterface because the user deleted the selected space. This happens here.

https://github.com/Underdisc/Varkor/blob/8293c4ad5fb98286967c19c48955cc4c3de86e5d/src/editor/CoreInterface.cc#L86

The CloseInterface function does not delete the interface though. It just sets a flag to mark it as deletable for the next time interfaces are handled. FindInterface now pays attention to that flag in addition to checking if the interface exists. This implies that if an interface's open flag is false, we should assume the data on that interface has been invalidated and is therefore no longer usable.