AssLib::TryGetLive now takes an optional default parameter and the default sprite shader has been added to the AssetLibrary.
The reason for the default parameter is to use a different default asset when the requested asset is invalid or fails to load. Text, for example, can use the default text shader instead of the basic default shader if the requested shader asset is invalid for some reason. There was an attempt to do this in the Renderer, but it disrupts the pop-in process.

As for the default sprite shader. I have no idea why it hasn't been added to the AssetLibrary yet.