Asset types are now initialized using InitInfo structs.
Before this, asset types were initialized with a vector of file paths. This worked as a general solution, but it's only going to become more crusty as time goes. Really, the set of parameters needed to initialize assets is unique to different types. What if we want to, for example, change the wrapping parameter for an image. At the moment this is hard coded, but now that there is an InitInfo struct, a parameter can be added to that struct to make that option configurable. The fact that all asset types use paths for initialization is just something they happen to have in common. Using only paths for initialization ignores all of those other possible parameters.

The code in AssetInterfaces.cc is very repitious because of what I just mentioned. All the asset types have file paths in their InitInfo structs and those paths need to be modifiable. The code is still specific to each asset type, though, and cosolidation would result in a headache that duplication does not.