The FileInterface now returns a path relative to res/.
There is some significant friction between whether the FileInterface should return an absolute or relative path. The existence of e9ffd95 and 7da7376 is evidence of this. For the time being, I'm going to go forward with the change in e9ffd95. My primary reason for this can be explained by the code within editor/AssetInterfaces.cc. This code sets the paths within the asset InitInfo structures and these paths are always relative. It's more work to correct an absolute path to a relative one rather than the other way around, so the FileInterface will give back a relative path.

The fact that I've gone back and forth on this might imply that the FileInterface should have an option for handing back relative or absolute paths. In fact, it should probably have an option to browse only res/ or the entire filesystem as well. There is no need for these things yet, though, so this change works for now.