Re-added the std::string template specialization for Vlk::Value::As.
This was removed in 423c41f because I thought it wasn't necessary. This is not the case, though. The template As function uses a string stream in order to extract a value. For some reason, extracting a string from a string stream will only extract up to the first space. The template specialization has been re-added to avoid this. Additionally, putting a string into a string stream to extract the same string is a bit ridiculous and it's nice to not be doing that anymore.

The vlk_Value unit test has also been modified to test this.