Function objects are now used to add Valkor extensions.
This is a continuation of fff6b2d. The code in that commit did not allow Valkor extensions to use partial template specializations because functions cannot use partial template specializations. This meant that extending Valkor with a templatized type was not possible unless the type was fully specialized. This commit addresses that issue. By using the new Serializer and Deserializer structs, it's possible for extensions to use partial template specializations.

Note: It is not possible to tell if an extended type can be fully deserialized. This means that using Explorer::As without a defaultValue argument could lead to an abort when used with an extended type. This is a hole in Explorer's interface, but there's also no immediate need to fill it.