Removed the lamda hell that made its way into World::Space a year ago.
I just got really fed up with looking at it. The lambdas removed in this commit are a perfect example of where lambdas are completely unnecessary. Not only that, but they cause problems too. Any crashes that happened inside of these lambdas had a large callstacks because of the lambdas. I'm happy to get rid of them, even if it's not the most effective use of my time.