Introduced the VEdit component function.
The Editor hooks are nice for situations where a component's editor interface needs to store some extra data. When this is not the case, which is most of the time, using the editor hook adds an annoying amount of boilerplate. So, if a component defines a VEdit function, it will be used for component inspection. The classic editor hook will be used if the function isn't defined.