Child references are now removed when a child with a parent is deleted.
This bug manifested as a crash in the Overview interface that was caused by deleting a child member. The parent's reference to that child wasn't removed and the Overview interface attempted to display that now invalid child.

The unit test for World::Space now checks for this as well. I also made some minor formatting changes to the relevant unit test to make it a bit easier to follow.