The distinction between required and nonrequired assets is now a distinction between serializable and nonserializable assets.
Originally, any nonrequired asset was serialized by the asset library. This does not account for creating a nonrequired asset that doesn't get serialized. For example, a user might want to specify an asset in their project code with something like this. `AssLib::Create<Gfx::Shader>("ShaderName", "VertexFile", "FragmentFile")`. Initially, there was no way to do this without this asset being serialized when saving 'assets.vlk'. This meant there would be two instances of this asset, one from the line that creates the asset and another from the deserialization of `assets.vlk`. By splitting up assets under the criteria of serializable and nonserializable instead of required and nonrequired, creating an asset like this without serializing it is now possible.

This also adds a AssetType::InitInfo::Prep function to asset types. This allows us the set the InitInfo of an asset without initializing it. That way we can defer initializing an asset until it is needed, which is something that any asset can take advantage of.