Added the VStaticInit component function.
A component might want to do some work that's applicable to all instances of that component once at the start of the program. For example, it might want to create an asset that all instances of that component will use. This specific example is why AssetLibrary::Init now happens before Registrar::Init.