Assimp is now compiled as part of the project.
The assimp libs and dlls have been removed in favor of this approach. This means that unit tests will no longer randomly require an assimp dll. Fixing is the main driver for this change.

Additionally, the standalone version of varkor now links against the varkor library. They were originally two targets with equivalent source files. The ext/ subdirectory was added once for each of the targets. This prevented adding the assimp subdirectory within ext/CMakeLists.txt because the same CMake configuration would be interpreted twice. With this new approach, ext/ only needs to be added once.

This also cleans the CMake scripts with some reordering and the use of more target specific commands where possible.