Introduced the VRender function.
This provides a way for components to render themselves without needing to integrate new code into the renderer. At the same time, it provides a way to keep the rendering code completely separate from other kinds of code within the components. This will be useful when a day comes that rendering code is unnecessary for a build (consider a server client).

The other components should eventually use this system and there needs to be a way to specify the order in which component types are rendered.