MSVC builds now use /permissive-
This ensures that the code conforms to the cpp standard without special msvc flavoring. This makes it possible to use the /std:c++20 flag and it should make migrating to other compilers easier.

Most of these changes involve replacing Util::Move with std::move. Util::Move was made for learning purposes and it wasn't valid with the /permissive- flag, so I decided to get rid of it. I still don't understand the full version of std::move, but it's not something I really care about right now.