Removed the old implementation for editor hooks.
This is largely a cleanup commit. This started because I wanted to remove Registrar::Register and use Type<Component>::Register directly instead to register components. This required removing editor hooks.

These editor hooks were first introduced to deal with the state needed to use the transform gizmo. At the time (82df71f), the transform editor hook contained all of the code needed for displaying the gizmo. This was changed in 7c9df33 and subsequently removed the need for editor hooks.

The hooks were kept because the Transform hook still had state used to control various gizmo options. This was annoying, though, because any time the hook was closed and reopened, those options would be reset to their default. This could be addressed by just making the values static, but that completely removes the need for editor hooks since those static values can just be put on the component itself. This commit does exactly that.

All component editing is now done through the VEdit function. This means hooks no longer have access to an Editor::Interface from which they were derived. This required me to remove Editor::SelectAssetInterface and replace it with a drag and drop solution. I didn't want to remove this at first, but I am now convinced this solution is better because Editor::SelectAssetInterface became more cumbersome to use as the number of assets increased.

VEdit also received an owner parameter because Comp::Transfrom::VEdit required it.