Gfx::Framebuffer no longer maintains a list of fullscreen framebuffers.
I don't know exactly why, but after basically any abort, another abort would occur because the Framebuffer destructor failed to remove its reference in smFullscreens. The method for resizing fullscreen framebuffers was not great, though, because it held onto pointers that couldn't be guaranteed to stick around.

So, to address the multiple issues, the fullscreen framebuffers are now explicitly resized in Gfx::Renderer when the window is resized.