TryGet* functions on Vlk::Value will never crash.
These are called TryGet* for a reason. They will attempt to get a Vlk::Value, but if it is not there for whatever reason, the attempt will fail and a nullptr will be returned.

I ran into a scenario where Vlk::Explorer called one of these TryGets and a crash occured because the Value::Type was incorrect. I fully expect a user to do the same and having it crash instead of just showing an error is very annoying.