Layers have been introduce to handle Space metadata.
World::Space contained a name and a cameraId. These shouldn't be there because it's not the Space's job to manage them. The Space should be responsible for managing members, the components on those members, and nothing more. I'd eventually like to get more out of space so it does just that job.