Gfx::Shader::InitInfo is now a vector of files.
The InitInfo before this contained strings to handle each of the different scheme types. It wasted some memory, but not too much. It just felt like a messy solution to allow both single and split shader sources.

This addresses that by making the whole InitInfo applicable to all shaders. This did require the addition of type headers in all shaders, but I don't believe the additional explicitness is a bad thing.

Originally, my plan was to replace the InitInfo structs with Vlk::Values. That would have addressed the memory problem as well. Unfortunately, using a Vlk::Value for InitInfo resulted in new messes. The largest being the lack of ability to know exactly what parameters the InitInfo is expected to have. This would have made writing the edit code for InitInfos exceptionally hacky.