Vlk::Value () operator will now find an existing pair or create a pair if an existing one isn't found.
This is the first commit for the asset refactor. This particular change allows Value's to be more flexible when it comes to modification. It provides an easier way to access already created pairs. Assets will make extensive use of this.

I took this as an opportunity to clean up the unit test and also flesh out a chunk of TryGet functions that are useful for exploring values.