Removed unnecessary move assignment from Ds::Vector::Insert.
Though the original implementation worked, the default construction has been replaced with a move construction. This only removes a single function call.

I also took this as an opportunity to clean up the unit tests for Insert. Insert was split into Insert0 and Insert1 to test the different conditions handled by the function.