Increase the breadth of debug infromation devlivered to stdout from crt.
Instead of only sending crt memory warnings to stdout, we now send errors and asserts as well. This gives more debugging information for memory issues. Buffer overrun is one example of the new debugging information.

Disabling leak output also undoes all of the work done by the enable. This was not the case beforehand.