Vlk::Parser mValueStack.Push(mValueStack.Top()) fix.
This statement has a serious issue. Top() returns a reference to the last element. Since Push takes an argument by reference, it takes that reference. If the mValueStack vector grows during the Push, the reference returned from Top is invalidated, therefore invalidating the reference given to push.

To get around this, we save a copy of the value stored at the top of mValueStack  That way we guarantee that no reference invalidation occurs.