Vlk::Value now supports moving, copying, and comparison.
The reason for this isn't contained in this commit. Vlk::Values are being used to store configurations for assets. When an asset Value is being modified through the editor, we need a way to know if it has been updated so we can update the relevant file when there is a change. We would otherwise need to write the file after every frame or introduce a lot of extra boilerplate to check for updates.

These new functions provide a much easier way to check for updates.