Generic ResourceType definitions.
This is one of the most fundamental elements of the new asset system. It's on the same level as resources themselves.

The goal of this is to reflect type information about resource types. This information then becomes accessible during runtime with an enum type Id. This makes it possible to do operations specific to the type without knowing what the original cpp type was. It's necessary in situations where we have to do an operation on a resource and we no longer have the cpp type available to us. This will be apparent in the commit that adds the new asset type, the type which will contain all of the resources.

Obvisiouly, Gfx::Material and World::Mold are not defined, but they will be in later commits.