Utitlity constructors/functions for ResourceId.
I am not enforcing a specific scheme when setting the id string directly. I considered it, but I don't know what would be effective. There needs to a be situation where not enforcing the contents of the id string causes a problem that could be avoided.

The comparison operation is being used in the code of the uncommited AssetInterface.