Initial implementation of refactored assets.
The intention is for assets to hold onto the memory being used by multiple resources. The model resource type, for example, will result in loading multiple resources of different types (meshes, materials, and images). The Asset is a way to keep all of those created resources in one "bucket". This makes finding where a specific resource came from easy.

To achieve this, the asset has one byte array allocation that all of the resources are stored in. The elements in this array are different types, one type for each resource. Tracking where these resources are within the byte array is achieved using resource descriptors.

This implementation still needs improvement. The byte array should be allocated using a vector to avoid allocation calls within the implementation. Additionally, all of the resource types need to follow a factory pattern so resource types share a unified interface for creation. For particularly large assets, a scheme for sorting resources will be necessary so they can be quickly fetched (only once the bottleneck becomes apparent, of course).