Initial commit of Gfx::UniformVector.
The goal of this is to simplify the interface for gpu uniforms. By storing uniform data in a UniformVector, setting the uniforms specified by materials, renderables, and renderable collections can be deferred until just before a render is actually performed. This allows components and projects to make use of custom uniforms without needing to call glUniform* functions directly or with other helper functions.

The usage of this structure will come in the later commits that introduce Gfx::Material and Gfx::Renderable.