Vlk::Explorer now provides a valid type function.
It is useful to make sure that the underlying Vlk::Value::Type is what the user expects. This addresses situations where it is more useful to inform the user about an incorrect Value::Type in an error message rather than ignoring it by using default values.

Consider Gfx::Font::Init(const Vlk::Explorer& configEx) as a simple example.