Merged Vlk::Serializer and Vlk::Deserializer into Vlk::Converter.
This puts all of the serialization and deserialization code into vlk/Value.*. The deserialization code used to be within Vlk::Explorer. Vlk::Value is deserializable though, so putting that functionality in Vlk::Explorer didn't make much sense. Besides that, combining to the two simplifies the interface for creating extensions.

This also adds Vlk::Value::As(const T& default). The default parameter used to only be possible with Vlk::Explorer. Now that Value is being used for configurations, it's useful to use As with a default Value because it will avoid crashing from an unitialized value or incorrect type. Take a look at Gfx::Cubemap::EditConfig(Vlk::Value* configValP) as an example.