Updated Gfx::Mesh for asset refactor.
Gfx::Mesh can now be initialized like other resources. If the mesh is loaded from file, that file must only contain a single mesh. This creates a differentiation between Gfx::Mesh and the updated Gfx::Model.

The code for converting an aiMesh into a Gfx::Mesh has been moved into an Init function. This code is also in Gfx::Model and will be removed in a later commit.