Vlk::Value::EnsureType added.
EnsureType and ExpectType take the places of ExpectType and HardExpectType, respectively. ExpectType did not allow values to be overwritten with a new type. Because of how useful this can be in scenarios where Vlk::Values are being edited and the underlying type is in flux, EnsureType is used in some areas where ExpectType was used previously. This can result in destructive actions, but the configurability advantage outweighs that, especially if good practices to avoid destructive actions are being followed.