Removed working/res/
This contained the components.vlk file. The purpose of components.vlk is for subprojects to be able to track what components are currently in the project. It might be added back to the Vakor repository in a later commit, but if it is, it will be placed inside of vres/.