Updated Gfx::Renderer for the Asset refactor.
The major thing to pay attention to here is the removal of rendering code for specific component types. The rendering code for text, sprite, skybox, and mesh components are no longer here because that is now handled by renderables.

This change also put the shadow data used in shaders into a uniform buffer. Unfortunately, the texture could not be placed into the uniform buffer because samplers are not allowed in uniform buffers.