Fixed the Math::Vector and Gfx::HdrColor deserialization extensions.
The previous versions of these functions called As without defaults. This would result in a crash if the As failed. Whether a crash can occur on a deserialization attempt should be clear in user code.

This also replaces the use of Vlk::Explorer in the deserialization functions with Vlk::Value. The deserialization functions do not need the features provided by the Explorer. It also makes the parameters the same as the other deserialization functions.