Update to Comp::ShadowMap for Asset refactor.
The code that actually renders the shadow map was moved to Gfx::Renderer in an earlier commit. Shadows make use of a uniform buffer and all uniform buffer work is handled in Gfx::Renderer. For the sake of keeping all of that uniform work within Gfx::Renderer, the shadow map render was moved to there. This is subject to change, though.