Added Comp::Mesh for rendering mesh resources.
Comp::Model used to take care of this job, but now there is a distinction between the two. Comp::Mesh uses a Mesh resource and Comp::Model uses a Model resource. Comp::Mesh only supports a single Mesh, while a Comp::Model supports the multiple Meshes and Materials in a Model.