Shader updates for Asset refactor.
These updates are necessary because of previous commits.

The vec2 aPos attribute is no longer supported. Everything is a vec3 now because it made the code simpler on the cpp side of things.

Some uniforms have changed names. These names will line up with the materials that will be added in a later commit.