Added Editor::AssetInterface.
This interface is used to display all of the resources that are part of an asset. The defined resources come from the asset file. The initialized resources are all of the resources that get created during the asset's initialization.

The Library interface is responsible for creating it, and, tbh, the interface isn't as elegant as I would like. This will be addressed in the future.