Added VResult alias for ValueResult.
This aims to shorten lines that use ValueResult without decreasing the readability of the lines. It can even be argued that just the existence of a template parameter is enough to know that a Result is a ValueResult, but I am keeping the V to avoid conflicts with the existing Result struct and for the sake of clarity.