Reorder of VResult constructor parameters.
Now the value part of the VResult constructor is always the first parameter. Not having this be this case resulted in some slightly smelly code when the value of a VResult was a string. Since the error is the first paramter, is the error the value or the Result? This is why the Rsl::ResolveProjPath and Rsl::ResolveResPath had an extra argument on their trailing return statements. Without it, the error would have been used as the value.

If a ValueResult contains an error, that error must be constructed in a Result before being given the VResult constructor. This is done is the same vein as the other reason for this change. Using a VResult like this is only necessary if the type of the value is not default constructable, though. Most of the time, you can just return a Result and that will be converted into a VResult.