Assets now prevent using the same resource name multiple times.
I had an asset that contained multiple different model resources. Upon initialization, these model resources create other resources whose names collide with the resources created by the other models. A check is now in place to prevent this.

This has performance implications depending on the asset and it might be better to avoid it entirely by not allowing multiple resource definitions in a single asset file.