Introduced Asset::smInitAsset.
This is just a pointer that references the asset currently being initialized. There are a few resource init functions that need to know what asset they are being initialized on because they create other resources that need to be added to the asset. smInitAsset allows us to avoid passing that asset as a parameter. This makes Gfx::Model::Init flush with the rest of the initialization functions.