Spaces are now serialized as ValueArrays.
I created a Space that used parent child relationships quite heavily. Some of the children shared their names with niblings. When the Space I created was serialized, a significant amount of information was lost because trying to create new pairs with already used member names as keys returned already existing pairs.

Change 1002dff introduced this behaviour to PairArrays. The problem is not that change, though. Pair keys should be strings that one would expect to see in the code that deals with the PairArray. They are used for access, not storage. So, to address the problem, Space Values are now ValueArrays with PairArray elements. The member name is stored in its associated PairArray element. I've wanted to make this change for a while now. This problem encouraged me to do so.

This also includes some minor cleanup changes.