Added a scale parameter for Model and Mesh resources.
This was introduced primarily because of fbx files. They are stored in cm. Despite assimp having an option to convert the fbx data from cm to m, it does nothing. Adding a scale parameter is another way to scale imported fbx modles from cm to m. It's also a generally useful parameter to have.

To be aware that using a single mesh from an imported fbx model will not undergo this scaling.