Replaced the LibraryInterface with a full file/resource browser.
The AssetInterface was very unintuitive to deal with. This is my attempt at addressing that problem. The goal is consolidate the FileInterface and the AssetInterface into a single interface that allows the users to browse through all of the files and assets in their project. Since the assets are synonymous to files within the vres/ and res/ directories, combining the two seems perfectly reasonable.

The initialization of the CoreInterface is wrapped within ImGui's start and end Frame calls because starting an ImGui frame initializes some state that the LibraryInterface constructor uses.

Be aware that changing an asset file while that asset is expanded in the LibraryInterface will result in a mismatch between the resources that are displayed and the resources that are in the file.