Assets are now ValueArrays.
This means that resources are now stored as PairArray elements within an asset ValueArray. This was done because resource names were stored as pair keys and couldn't be modified. Now they are stored as TrueValues within a resource's PairArray which means the editor can be given the ability to change resource names.

This also includes some other minor cleanup work.