Assets are no longer collected during VarkorInit.
Collecting assets during VarkorInit filled up the asset tree with all of the assets in a project even though a majority of them would sit in the tree as dormant assets.

Now assets are only added to the asset tree once they are needed. This has extended the meaning of the dormant asset status. It can mean that the asset is within the tree but has not been queued for initialization, or it is not in the tree at all.