An option for adding resource.
I removed the initializing check before writing a resource config because it can potentially throw away changes. To make matters worse, the problem is larger than the single if check made it out to be. If the shared config value changes while the asset is initializing, race conditions can occur.