Different build types are easier to specify and no longer overwrite each other.
Using different build types was annoying. If you wanted to switch from debug to release, you would need to change that in buildSpecs.bat. Now the build type is an argument to build.bat and btest.bat, making it easier to specify and change.

Additionally, the output binaries are now postfix with the type of build they are. This makes it obvious whether a binary is debug, release, etc. and it stops those binaries from being thrown away when a binary with a different build type is built.