Vlk::Value can now be assigned to with pairs.
You'd think that this isn't required. Vlk::Pair inherits from Vlk::Value, so it seems like assigning a Pair to a Value should just call the assignment function that takes a Value argument. 63c117ef added a templatized assignemt operator to Vlk::Value. Because of this, assigning a Pair to a Value calls this template function rather than the desired assignment operator. Now we've explicityly added the assignment operator.