Updated golden file.
The vlk_Extenions golden file was out of date during some commit where deserialization was being worked on. This invalidation was reasonable though. It does not make sense to print out an error when a default value is provided to the As function. We should quitely use the default if serialization fails because that's what the user asked for. We will likely need to add another function that allows the user to check the status of deserialization on their own if they want to know about the deserialization status.