Build outputs are no longer placed in working.
With the addition of dealing with different build types and now clang, the number of executables placed in the working directory has gotten unreasonable. To address this, all build outputs are kept within the build dir. When the user requests to run the output from the build script, the file is moved into working, ran, and moved back to the directory it came from.