Component names no longer use typeid.
The names given by typeid are different when compiling on linux. It doesn't mean typeid is entirely useless (check out Interface.hh), but it does not fit the requirement for component names. It should also be noted that it would be harder to extract the desired string from the typeid name.