Threaded asset loading bug fix.
On some runs in release builds, the init thread would never join and this would prevent one's ability to close the program. This was because an asset was added to the init queue just after the init thread exited its main initialization loop, but also just before the main thread checked whether the init queue was empty so it could join the thread. Since the init queue had something in it and the init thread wasn't joined, the init thread would never join and get recreated.