Fixed a segfault under g++.
When the Asset tree is cleared after exiting main on g++, a segfault occurs when calling one of the OpenGL cleanup functions. It was `glDeleteTextures` in my test case. By clearing the asset tree within Rsl::Purge, we can avoid the segfault.

For some reason this segfault does not show up on other compilers. This is surprising because glfwTerminate() is called in any case. It seems reasonable that gl functions are no longer available no matter which compiler is used.