Integrated tracy and added some tracy marks.
Tracy will only work under a RelWithDepInfo configuration. I don't want it included in Release or MinSizeRel builds. It's also not as useuful in Debug builds, understandbly. wolfpld even mentions this in the Tracy documentation. That leaves us with RelWithDebInfo. Considering profiling information is debug information, it feels very natural to only include Tracy with this configuration.