Abort on GlError::Callback.
Just logging the error and printing a callstack was not helpful. The result was spamming the console window with the opengl errors and stack traces to the point where the original error gets burried.

This will need to be adjusted if GL_DEBUG_TYPE_PERFORMANCE or GL_DEBUG_SEVERITY_NOTIFICATION messages are given to the callback again.