Use different default resource with TryGetRes.
The resource to be treated as the default is highly context dependent. Consider two cases: rendering a world object and rendering a fullscreen mesh. If we are rendering a world object, we need to use a shader that will apply the model, view, and projection matrices to the mesh vertices. If we are rendering a fullscreen mesh, we don't want the default shader to apply transformation matrices since the mesh being rendered shouldn't need any. So the default shader we want depends on what the the thing we are rendering is and is therefore context dependant.