Bloom
The implemenation involves a separated gaussian blur. The number of blurs is configurable, but defaults to 3.

There is no longer a framebuffer for every layer. Layers should instead be blended directly after they are done being rendered to and cleared before they are rendered to. The old system seemed like a waste of time (maintanence wise) and space (memory wise).

This also resulted in a change to post processing / tone mapping. Here's a comparison of the old and new system.

- Old -
Render renderables to LayerBuffer.
Post process and/or tonemap LayerBuffer into default Framebuffer.

- New -
Render renderables to LayerBuffer.
Post process and blend LayerBuffer into BlendBuffer.
Additive blend BlendBuffer and blurred IntenseBuffer into FinalBuffer.
Tonemap FinalBuffer into default frambuffer.

The major takeaway is that there is a seperate step for "post processing" and tonemapping. The idea is to give the user control over any post processing effects that are applied to a layer without having to worry about tonemapping. This is certainly a case of over-engineering.