Reorganized camera matrices.
The renderer was passing around the projection and view matrices to functions as multiple parameters. It's nice to package up that data into a single camera object that contains the camera component.

Since the editor camera uses a transform and camera component under the hood, the editor camera is now an object in the editor space. Since the editor camera is an object, it can be used in the renderer in the same way as any other camera.

This also shifts all the editor camera functionality (besides editor controls) to the camera component. It's not only applicable to the editor camera.